requirement.txt
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv
uvicorn
pydantic
httpx
langchain_google_community
langchain_tavily
langchain_groq
langchain_openai
yarl
langgraph 
langchain-google-community[places]


-e

